% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tabset_codes.R
\name{add_tabset_codes}
\alias{add_tabset_codes}
\title{Inserts Tabset Syntax by Language}
\usage{
add_tabset_codes(chunk_label)
}
\arguments{
\item{chunk_label}{chunk label stub}
}
\description{
Inserts Tabset Syntax by Language
}
